refactor: move restore all plugins
