+++ packages/mask/background/services/backup/internal_restore.ts
@@ -1,5 +1,6 @@
 import { delay } from '@dimensiondev/kit'
 import type { NormalizedBackup } from '@masknet/backup-format'
+import { activatedPluginsWorker, registeredPluginIDs } from '@masknet/plugin-infra/background-worker'
 import { ProfileIdentifier, RelationFavor } from '@masknet/shared-base'
 import { MaskMessages } from '../../../shared/messages'
 import {
@@ -14,13 +15,9 @@ import type { LatestRecipientDetailDB, LatestRecipientReasonDB } from '../../dat
 
 // Well, this is a bit of a hack, because we have not move those two parts into this project yet.
 // TODO: MV3 support
-let restorePlugins: (backup: NormalizedBackup.Data['plugins']) => Promise<void>
 let restoreWallets: (backup: NormalizedBackup.WalletBackup[]) => Promise<void>
 export function delegateWalletRestore(f: typeof restoreWallets) {
     restoreWallets = f
-}
-export function delegatePluginRestore(f: typeof restorePlugins) {
-    restorePlugins = f
 }
 export async function restoreNormalizedBackup(backup: NormalizedBackup.Data) {
     const { plugins, posts, wallets } = backup
@@ -178,4 +175,35 @@ function restorePosts(backup: Iterable<NormalizedBackup.PostBackup>) {
         }
         await Promise.all(promises)
     })
+}
+async function restorePlugins(backup: NormalizedBackup.Data['plugins']) {
+    const plugins = [...activatedPluginsWorker]
+    const works = new Set<Promise<void>>()
+    for (const [pluginID, item] of Object.entries(backup)) {
+        const plugin = plugins.find((x) => x.ID === pluginID)
+        // should we warn user here?
+        if (!plugin) {
+            if ([...registeredPluginIDs].includes(pluginID))
+                console.warn(`[@masknet/plugin-infra] Found a backup of a not enabled plugin ${plugin}`, item)
+            else console.warn(`[@masknet/plugin-infra] Found an unknown plugin backup of ${plugin}`, item)
+            continue
+        }
+
+        const f = plugin.backup?.onRestore
+        if (!f) {
+            console.warn(
+                `[@masknet/plugin-infra] Found a backup of plugin ${plugin} but it did not register a onRestore callback.`,
+                item,
+            )
+            continue
+        }
+        works.add(
+            (async () => {
+                const x = await f(item)
+                if (x.err) console.error(`[@masknet/plugin-infra] Plugin ${plugin} failed to restore its backup.`, item)
+                return x.unwrap()
+            })(),
+        )
+    }
+    await Promise.allSettled(works)
 }
